- Changed: HIRESTEX 'define' textures now replace existing textures   of type MiscPatch with the same name. - Added UDMF line trigger types MonsterUse and MonsterPush. - Separated skill and class filter bits from FMapThing::flags so that   UDMF can define up to 16 of each. Also separated easy/baby and   hard/nightmare and changed default MAPINFO definitions. - Changed: Crosshair drawing uses the current player class's default health instead   of 100 to calculate the color for the crosshair. - Added SECF_NOFALLINGDAMAGE flag plus Sector_ChangeFlags to set it. Also separated   all user settable flags from MoreFlags into their own Flags variable.
SVN r964 (trunk)

+++ src/decallib.h
@@ -61,7 +61,7 @@ protected:
 
 	FDecalBase *Left, *Right;
 	FName Name;
-	BYTE SpawnID;
+	WORD SpawnID;
 	TArray<const PClass *> Users;	// Which actors generate this decal
 };
 
@@ -98,22 +98,22 @@ public:
 	void ReadDecals (FScanner &sc);
 	void ReadAllDecals ();
 
-	const FDecalTemplate *GetDecalByNum (BYTE num) const;
+	const FDecalTemplate *GetDecalByNum (WORD num) const;
 	const FDecalTemplate *GetDecalByName (const char *name) const;
 
 private:
 	struct FTranslation;
 
 	static void DelTree (FDecalBase *root);
-	static FDecalBase *ScanTreeForNum (const BYTE num, FDecalBase *root);
+	static FDecalBase *ScanTreeForNum (const WORD num, FDecalBase *root);
 	static FDecalBase *ScanTreeForName (const char *name, FDecalBase *root);
 	static void ReplaceDecalRef (FDecalBase *from, FDecalBase *to, FDecalBase *root);
 	FTranslation *GenerateTranslation (DWORD start, DWORD end);
-	void AddDecal (const char *name, BYTE num, const FDecalTemplate &decal);
+	void AddDecal (const char *name, WORD num, const FDecalTemplate &decal);
 	void AddDecal (FDecalBase *decal);
 	FDecalAnimator *FindAnimator (const char *name);
 
-	BYTE GetDecalID (FScanner &sc);
+	WORD GetDecalID (FScanner &sc);
 	void ParseDecal (FScanner &sc);
 	void ParseDecalGroup (FScanner &sc);
 	void ParseGenerator (FScanner &sc);
